finish refactoring action endpoint
