+++ app/interactivity/interactions.ts
@@ -1,7 +1,3 @@
-export enum InitiativeCallbackAction {
-  EDIT_INITIATIVE_DIALOG = 'EDIT_INITIATIVE_DIALOG'
-}
-
 export enum InitiativeAction {
   JOIN_AS_MEMBER = 'JOIN_AS_MEMBER',
   JOIN_AS_CHAMPION = 'JOIN_AS_CHAMPION',
@@ -9,7 +5,12 @@ export enum InitiativeAction {
   UPDATE_STATUS = 'UPDATE_STATUS',
   DELETE = 'DELETE',
   OPEN_EDIT_DIALOG = 'OPEN_EDIT_DIALOG',
-  UPDATE_INITIATIVE = 'UPDATE_INITIATIVE'
+  EDIT_INITIATIVE = 'EDIT_INITIATIVE',
+  MODIFY_INITIATIVE = 'MODIFY_INITIATIVE',
+  MARK_ACTIVE = 'MARK_ACTIVE',
+  MARK_ON_HOLD = 'MARK_ON_HOLD',
+  MARK_COMPLETE = 'MARK_COMPLETE',
+  MARK_ABANDONED = 'MARK_ABANDONED'
 }
 
 export enum MemberAction {
@@ -18,12 +19,5 @@ export enum MemberAction {
   REMOVE_MEMBER = 'REMOVE_MEMBER',
   REMAIN_MEMBER = 'REMAIN_MEMBER',
   OPEN_EDIT_DIALOG = 'OPEN_EDIT_DIALOG',
-  UPDATE_ROLE = 'UPDATE_ROLE'
-}
-
-export enum StatusUpdateAction {
-  MARK_ACTIVE = 'MARK_ACTIVE',
-  MARK_ON_HOLD = 'MARK_ON_HOLD',
-  MARK_COMPLETE = 'MARK_COMPLETE',
-  MARK_ABANDONED = 'MARK_ABANDONED'
+  MODIFY_MEMBER = 'MODIFY_MEMBER'
 }
