+++ app/shared/initiatives-table.ts
@@ -1,0 +1,3 @@
+import { DynamoDB } from 'aws-sdk';
+
+export const initiativesTable = new DynamoDB.DocumentClient({ region: process.env.REGION, apiVersion: '2012-08-10' });
